requirements.txt
annotated-types==0.6.0
anyio==4.3.0
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pdfminer.six==20231228
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
ruff==0.3.4
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tabula-py==2.9.0
typing_extensions==4.10.0
tzdata==2024.1
